+++ internal/addoninstancesdk/addoninstance_watcher.go
@@ -1,4 +1,4 @@
-package utils
+package addoninstancesdk
 
 import (
 	"context"
